src/ml_pipeline/requirements.txt
google-cloud-storage==2.5.0
google-cloud-secret-manager==2.12.5
mlflow==1.29.0
